Feat: added slice - "resumes", types, and reducers
The slice is "resumes", in which the functionality of adding a different data user who
creates a resume will be performed. At the moment, it has the following
editors: "setNameResume" - adding the name of the resume, which will later
be found in the search for developers; "setBasicInfo" - adding contact information
for further cooperation with the developer: first name, last name, e-mail, as well
as work permit and citizenship; "setEducationClass" - a reduxer, which allows you
to fill in the data of the degree of education in information systems,
if any: degree, name of institution, year of graduation; "deleteResume" -
deleting the resume. In the future, other editors will be created to create
a developer's resume, such as: "setChangeTypeWork" - the category
of developer experience, that is, the applicant can choose the following category -
[described in the typesResume file.ts, line 69-71]; "SetPositions", "setProjects" -
reducers that allow you to perform the functionality of adding the company where the
applicant worked, its where the applicant worked, extensive information about her;
"setProjects" - has similar functionality, but stores data about created favorite
projects. "setSkils" - adds various skills to job seekers, there
are various recommended skills and a huge list to search for.